Makes mailboxed more robust.
Currently, the `Ord` instance on `mailboxed` is not used to its fullest potential, as we never know the key stored in the mailbox. This means that if a type has a novel `Eq` implementation, we won't know what it was equal to.

This is particularly useful in cases where we have an internal ID and we receive a server-side ID. In that case, we can have:

data InternalId = InternalId ClientId ServerId

instance Eq InternalId where
  eq (InternalId cid0 _) (InternalId cid1 _) = cid0 == cid1

Equality is only defined in terms of the client side id, and when mailboxed is used, we can listen for the client-side ID to get the server side id.