Longish commit message about keepLatest.
The problem with all of the previous approaches is evident when you write it out
by hand.

keepLatestHack :: forall a. Poll (Poll a) -> Poll a
keepLatestHack a = APoll \e -> makeLemmingEvent \s k ->
  s (sample_ a e) \p -> let _ = spy "toplevel works" true in void $ s (sample p e) \i -> do
    let _ = spy "coming through" i
    k i

`sample_ a e` emits, but by the time it has emitted, it means that whatever is
_INSIDE OF_ `a` has emitted with is either at the same time as OR after e.
If it is after `e`, then there is no way the internal sample will go through
because `e` isn't scheduled to emit again. What we really need is a fixed point
over the outer event and then to once it in the inner event!